@@ -34,11 +34,11 @@ JSONSchemaItem = TypedDict(
         "writeOnly": bool,
         "examples": list[Any],
         # exclusiveMinimum: 0
-        "multipleOf": Union[int, float],
-        "maximum": Union[int, float],
-        "exclusiveMaximum": Union[int, float],
-        "minimum": Union[int, float],
-        "exclusiveMinimum": Union[int, float],
+        "multipleOf": int | float,
+        "maximum": int | float,
+        "exclusiveMaximum": int | float,
+        "minimum": int | float,
+        "exclusiveMinimum": int | float,
         # minimum: 0
         "maxLength": "_NonNegativeInteger",
         # minimum: 0
